requirements.txt
streamlit
yt-dlp
ffmpeg-python
numpy